move to float matrix weightings

+++ db/migrate/20220525024458_create_transfer_layer.rb
@@ -4,7 +4,7 @@ class CreateTransferLayer < ActiveRecord::Migration[6.1]
   def change
     create_table :transfer_layers do |t|
       t.references :bot, null: false
-      t.integer :layer_matrix, array: true, null: false
+      t.float :layer_matrix, array: true, null: false
       t.integer :depth, null: false
       t.integer :row_count, null: false
       t.integer :col_count, null: false
