+++ db/schema.rb
@@ -33,7 +33,7 @@ ActiveRecord::Schema.define(version: 2022_05_25_024458) do
 
   create_table "transfer_layers", force: :cascade do |t|
     t.bigint "bot_id", null: false
-    t.integer "layer_matrix", null: false, array: true
+    t.float "layer_matrix", null: false, array: true
     t.integer "depth", null: false
     t.integer "row_count", null: false
     t.integer "col_count", null: false
